@@ -1,124 +1,199 @@
 {
   "name": "openfin-react-starter",
-  "version": "0.10.5",
+  "version": "0.30.10",
   "private": true,
   "dependencies": {
-    "@albertli/openfin-browser-adapter": "^0.10.5",
-    "@albertli/redux-openfin": "^0.10.5",
-    "@material-ui/core": "^1.5.0",
-    "@material-ui/icons": "^2.0.2",
+    "chalk": "2.4.1",
+    "dotenv": "6.0.0",
+    "dotenv-expand": "4.2.0",
+    "express": "^4.16.4",
+    "hadouken-js-adapter": "^0.38.1",
+    "morgan": "^1.9.0",
+    "rimraf": "^2.6.2",
+    "tcp-port-used": "^1.0.0"
+  },
+  "devDependencies": {
+    "@albertli90/openfin-browser-adapter": "^0.30.6-beta",
+    "@albertli90/redux-openfin": "^0.30.5-beta",
+    "@babel/core": "7.1.2",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/plugin-proposal-decorators": "^7.1.2",
+    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-transform-classes": "^7.1.0",
+    "@babel/plugin-transform-destructuring": "^7.1.3",
+    "@babel/plugin-transform-modules-commonjs": "^7.1.0",
+    "@babel/plugin-transform-react-constant-elements": "^7.0.0",
+    "@babel/plugin-transform-react-display-name": "^7.0.0",
+    "@babel/plugin-transform-regenerator": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.1.0",
+    "@babel/preset-env": "^7.1.0",
+    "@babel/preset-flow": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "^7.1.0",
+    "@babel/runtime": "7.1.2",
+    "@material-ui/core": "^3.6.1",
+    "@material-ui/icons": "^3.0.1",
+    "@svgr/webpack": "2.4.1",
     "@types/classnames": "^2.2.6",
     "@types/enzyme": "^3.1.13",
+    "@types/enzyme-adapter-react-16": "^1.0.3",
     "@types/history": "^4.7.0",
     "@types/jest": "^23.3.1",
     "@types/node": "^10.7.0",
-    "@types/react": "^16.4.10",
+    "@types/react": "^16.4.13",
     "@types/react-dom": "^16.0.7",
-    "@types/react-redux": "^6.0.6",
     "@types/react-router-dom": "^4.3.0",
     "@types/redux-actions": "^2.3.0",
-    "autoprefixer": "7.1.6",
-    "babel-core": "6.26.0",
-    "babel-eslint": "7.2.3",
-    "babel-jest": "20.0.3",
-    "babel-loader": "7.1.2",
-    "babel-preset-react-app": "^3.1.1",
-    "babel-runtime": "6.26.0",
-    "case-sensitive-paths-webpack-plugin": "2.1.1",
-    "chalk": "1.1.3",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "9.0.0",
+    "babel-jest": "23.6.0",
+    "babel-loader": "8.0.4",
+    "babel-plugin-named-asset-import": "^0.2.3",
+    "babel-preset-react-app": "^6.1.0",
+    "bfj": "6.1.1",
+    "case-sensitive-paths-webpack-plugin": "2.1.2",
     "classnames": "^2.2.6",
     "color": "^3.0.0",
-    "css-loader": "0.28.7",
-    "dotenv": "4.0.0",
-    "dotenv-expand": "4.2.0",
-    "enzyme": "^3.4.1",
-    "enzyme-adapter-react-16": "^1.2.0",
-    "eslint": "4.10.0",
-    "eslint-config-react-app": "^2.1.0",
-    "eslint-loader": "1.9.0",
-    "eslint-plugin-flowtype": "2.39.1",
-    "eslint-plugin-import": "2.8.0",
-    "eslint-plugin-jsx-a11y": "5.1.1",
-    "eslint-plugin-react": "7.4.0",
-    "express": "^4.16.3",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.5",
-    "fs-extra": "3.0.1",
+    "css-loader": "1.0.0",
+    "dexie": "^2.0.4",
+    "enzyme": "^3.7.0",
+    "enzyme-adapter-react-16": "^1.6.0",
+    "enzyme-to-json": "^3.3.5",
+    "eslint": "5.6.0",
+    "eslint-config-react-app": "^3.0.5",
+    "eslint-loader": "2.1.1",
+    "eslint-plugin-flowtype": "2.50.1",
+    "eslint-plugin-import": "2.14.0",
+    "eslint-plugin-jsx-a11y": "6.1.2",
+    "eslint-plugin-react": "7.11.1",
+    "fake-indexeddb": "^2.0.4",
+    "file-loader": "2.0.0",
+    "fork-ts-checker-webpack-plugin": "^0.4.14",
+    "fs-extra": "^7.0.1",
     "history": "^4.7.2",
-    "html-webpack-plugin": "2.29.0",
-    "jest": "21.2.1",
-    "morgan": "^1.9.0",
+    "html-webpack-plugin": "4.0.0-alpha.2",
+    "identity-obj-proxy": "3.0.0",
+    "jest": "23.6.0",
+    "jest-environment-jsdom": "^23.4.0",
+    "jest-environment-jsdom-global": "^1.1.0",
+    "jest-pnp-resolver": "1.0.1",
+    "jest-resolve": "23.6.0",
+    "loader-utils": "^1.1.0",
+    "material-ui-pickers": "^1.0.1",
+    "mini-css-extract-plugin": "0.4.3",
+    "moment": "^2.22.2",
     "object-assign": "4.1.1",
-    "openfin-launcher": "^1.4.4",
-    "postcss-flexbugs-fixes": "3.2.0",
-    "postcss-loader": "2.0.8",
+    "optimize-css-assets-webpack-plugin": "5.0.1",
+    "pkg": "^4.3.5",
+    "pnp-webpack-plugin": "1.1.0",
+    "postcss-flexbugs-fixes": "4.1.0",
+    "postcss-loader": "3.0.0",
+    "postcss-preset-env": "6.0.6",
+    "postcss-safe-parser": "4.0.1",
     "promise": "8.0.1",
     "raf": "3.4.0",
-    "react": "^16.4.2",
+    "react": "^16.6.3",
+    "react-app-polyfill": "^0.1.3",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dev-utils": "^5.0.1",
-    "react-dom": "^16.4.2",
+    "react-dev-utils": "^6.1.1",
+    "react-dom": "^16.6.3",
     "react-json-view": "^1.19.1",
+    "react-number-format": "^3.6.0",
     "react-particles-js": "^2.3.0",
     "react-redux": "^5.0.7",
     "react-router-dom": "^4.3.1",
     "react-splitter-layout": "^3.0.1",
     "redux": "^4.0.0",
     "redux-actions": "^2.6.1",
+    "redux-mock-store": "^1.5.3",
     "redux-saga": "^0.16.0",
-    "resolve": "1.6.0",
-    "rimraf": "^2.6.2",
-    "source-map-loader": "^0.2.3",
-    "style-loader": "0.19.0",
-    "sw-precache-webpack-plugin": "0.11.4",
-    "ts-jest": "^23.1.3",
-    "ts-loader": "^3.5.0",
-    "typescript": "^3.0.1",
-    "url-loader": "0.6.2",
-    "webpack": "3.8.1",
-    "webpack-dev-server": "2.9.4",
-    "webpack-manifest-plugin": "1.3.2",
-    "whatwg-fetch": "2.0.3"
+    "redux-saga-test-plan": "^3.7.0",
+    "resolve": "1.8.1",
+    "sass-loader": "7.1.0",
+    "shelljs": "^0.8.3",
+    "shortid": "^2.2.13",
+    "style-loader": "0.23.0",
+    "terser-webpack-plugin": "1.1.0",
+    "thread-loader": "1.1.5",
+    "ts-jest": "^23.1.4",
+    "tslint": "^5.11.0",
+    "tslint-config-prettier": "^1.15.0",
+    "tslint-react": "^3.6.0",
+    "typescript": "^3.0.3",
+    "url-loader": "1.1.1",
+    "webpack": "4.19.1",
+    "webpack-dev-server": "3.1.9",
+    "webpack-manifest-plugin": "2.0.4",
+    "whatwg-fetch": "2.0.4",
+    "workbox-webpack-plugin": "3.6.3"
   },
   "scripts": {
     "start": "node scripts/start.js",
     "start-openfin": "node scripts/start.openfin.js",
-    "build": "rimraf build && node scripts/build.js",
-    "server": "node scripts/server.js",
-    "server-openfin": "node scripts/server.openfin.js",
-    "test": "node scripts/test.js --env=jsdom"
+    "build": "rimraf build && rimraf package && node scripts/build.js",
+    "serve": "node scripts/server.js",
+    "serve-openfin": "node scripts/server.openfin.js",
+    "package": "node scripts/package.js",
+    "test": "node scripts/test.js"
+  },
+  "browserslist": {
+    "development": [
+      "last 2 chrome versions",
+      "last 2 firefox versions",
+      "last 2 edge versions"
+    ],
+    "production": [
+      ">0.25%",
+      "not op_mini all",
+      "ie 11"
+    ]
   },
   "jest": {
     "collectCoverageFrom": [
-      "src/**/*.{ts,tsx,js,jsx,mjs}"
+      "src/**/*.{ts,tsx,js,jsx,mjs}",
+      "!src/**/*.d.ts",
+      "!src/serviceWorker.js",
+      "!src/index.tsx"
     ],
     "setupFiles": [
       "<rootDir>/config/polyfills.js",
       "<rootDir>/config/jest/setupTests.js"
     ],
     "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.{ts,tsx,js,jsx,mjs}",
-      "<rootDir>/src/**/?(*.)(spec|test).{ts,tsx,js,jsx,mjs}"
+      "**/__tests__/**/*.{ts,tsx,js,jsx,mjs}",
+      "**/?(*.)(spec|test).{ts,tsx,js,jsx,mjs}"
     ],
-    "testEnvironment": "jsdom",
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ],
+    "roots": [
+      "<rootDir>/src"
+    ],
+    "globals": {},
+    "testEnvironment": "jest-environment-jsdom-global",
     "testURL": "http://localhost",
     "transform": {
-      "^.+\\.(tsx|ts)$": "ts-jest",
-      "^.+\\.(js|jsx|mjs)$": "<rootDir>/node_modules/babel-jest",
+      "^.+\\.tsx?$": "ts-jest",
+      "^.+\\.(ts|tsx|js|jsx|mjs)$": "<rootDir>/config/jest/babelTransform.js",
       "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|mjs|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
+      "^(?!.*\\.(js|jsx|ts|tsx|mjs|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
     },
     "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs)$"
+      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs|ts|tsx)$",
+      "^.+\\.module\\.(css|sass|scss)$"
     ],
     "moduleNameMapper": {
-      "^react-native$": "react-native-web"
+      "^react-native$": "react-native-web",
+      "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
     },
     "moduleFileExtensions": [
-      "ts",
-      "tsx",
       "web.js",
       "js",
+      "web.ts",
+      "ts",
+      "web.tsx",
+      "tsx",
       "json",
       "web.jsx",
       "jsx",
