@@ -1,35 +1,54 @@
 import * as React from 'react';
-import * as cx from 'classnames';
+import cx from 'classnames';
 
-import {withStyles} from '@material-ui/core/styles';
+import { WithStyles, withStyles} from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
-import IconButton from '@material-ui/core/IconButton';
-import Hidden from '@material-ui/core/Hidden';
 import Button from '@material-ui/core/Button';
+import Fab from '@material-ui/core/Fab';
 
 import Menu from '@material-ui/icons/Menu';
 import ChevronLeft from '@material-ui/icons/ChevronLeft';
 
 import HeaderLinks from './HeaderLinks';
 
+import {MouseEventHandler} from "react";
+
 import { headerCompStyle as style } from '../../assets/jss/openfin-starter';
+import {RouteItem} from '../../routes';
 
-const appLogo = require('../../assets/svg/app.svg') as string;
+import appLogo from'../../assets/svg/app.svg';
 
-class HeaderComp extends React.Component<any,any>{
+interface IProps extends WithStyles<any> {
+    routes:RouteItem[],
+    color:string,
+    open?:boolean,
+    windowsState:string,
+    docked?:boolean,
+    handleDrawerToggle?:MouseEventHandler<any>,
+    onSwitchToLaunchBar?:MouseEventHandler<any>,
+    onUndock?:MouseEventHandler<any>,
+    onMinimize:MouseEventHandler<any>,
+    onMaximize:MouseEventHandler<any>,
+    onClose:MouseEventHandler<any>,
+    // todo: maybe should use,  import {RouteComponentProps} from "react-router"; but feel like it did not work the moment i tried......
+    [key:string]:any,
+    [key:number]:any,
+}
+
+class HeaderComp extends React.Component<IProps,{}>{
 
     render(){
 
         const {
-            classes, color, windowState,
-            handleDrawerToggle,
-            onSwitchToLaunchBar, onMinimize, onMaximize, onClose,
+            classes, color, windowsState,
+            handleDrawerToggle, docked,
+            onSwitchToLaunchBar, onUndock, onMinimize, onMaximize, onClose,
         } = this.props;
 
         function makeBrand(props:any) {
             let name = null;
-            props.routes.forEach((prop:any, key:number):any => {
+            props.routes.forEach((prop:RouteItem):any => {
                 if (prop.path === props.location.pathname){
                     name = prop.navbarName;
                 }
@@ -41,24 +60,25 @@ class HeaderComp extends React.Component<any,any>{
             <Toolbar className={classes.container}>
                 <div className={classes.flex}>
                     {handleDrawerToggle?
-                        <Button
+                        <Fab
                             className={classes.menuBtn}
-                            variant={"fab"}
-                            mini
                             color={"inherit"}
                             aria-label={"open drawer"}
                             onClick={handleDrawerToggle}
                         >
-                            {this.props.open?<ChevronLeft/>:<Menu/>}
-                    </Button>:null}
+                                {this.props.open?<ChevronLeft/>:<Menu/>}
+                        </Fab>
+                    :null}
                     <img src={appLogo} className={classes.companyLogImg}/>
                     <Button href={"#"} className={classes.title}>
                         {"App name| "+ makeBrand(this.props)}
                     </Button>
                 </div>
                 <HeaderLinks
-                    windowState={windowState}
+                    windowsState={windowsState}
+                    docked = {docked}
                     onSwitchToLaunchBar={onSwitchToLaunchBar}
+                    onUndock={onUndock}
                     onMinimize={onMinimize}
                     onMaximize={onMaximize}
                     onClose={onClose}
