@@ -1,15 +1,15 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
-import * as cx from 'classnames';
+import cx from 'classnames';
 import { Scrollbars } from 'react-custom-scrollbars';
 
-import { withStyles, StyleRules } from '@material-ui/core/styles';
+import { WithStyles, withStyles } from '@material-ui/core/styles';
 
-import { Window } from '@albertli/redux-openfin';
+import { Window } from '@albertli90/redux-openfin';
 
 import AppBar from '@material-ui/core/AppBar';
 import Toolbar from '@material-ui/core/Toolbar';
-import Button from '@material-ui/core/Button';
+import Fab from '@material-ui/core/Fab';
 import IconButton from '@material-ui/core/IconButton';
 
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
@@ -20,15 +20,29 @@ import ZoomOutMapIcon from '@material-ui/icons/ZoomOutMap';
 
 import { launchBarLayoutStyle as style } from '../../assets/jss/openfin-starter';
 
-const appLogo = require('../../assets/svg/app.svg') as string;
+import appLogo from '../../assets/svg/app.svg';
 
 import {
-    applicationLaunchBarToggle, applicationLaunchBarToggleCollapse
-} from '../../redux';
+    // actions
+    applicationLaunchBarToggle, applicationLaunchBarToggleCollapse, applicationLaunchNewWindow,
+    // types
+    IRootState,
+} from '../../reduxs';
 
 import { launchBarItems } from './LaunchBarData';
 
-class LaunchBarComp extends React.Component<any,any>{
+interface IProps extends WithStyles<typeof style>{
+    launchBarCollapse:boolean,
+    actions:{
+        handleLaunchBarItemBtnClick: (appJson:any)=>()=>void,
+        handleSwitchToMainWindow: ()=>void,
+        handleToggleCollapse: ()=>void,
+        handleMinimize: ()=>void,
+        handleClose: ()=>void,
+    }
+}
+
+class LaunchBarComp extends React.Component<IProps,{}>{
     render(){
 
         const {
@@ -44,9 +58,9 @@ class LaunchBarComp extends React.Component<any,any>{
 
         const collapse = launchBarCollapse;
 
-        const _launchBarItems = launchBarItems;
+        // const _launchBarItems = launchBarItems;
 
-        const buttonContainerWidth = _launchBarItems.length<10?_launchBarItems.length*64:576;
+        const buttonContainerWidth = launchBarItems.length<10?launchBarItems.length*64:576;
 
         return (
             <span>
@@ -64,7 +78,7 @@ class LaunchBarComp extends React.Component<any,any>{
                                 width:`${buttonContainerWidth}px`,
                             }}
                         >
-                            {_launchBarItems.map((item,index)=>{
+                            {launchBarItems.map((item,index)=>{
                                 if (item.icon){
                                     return <IconButton key={index}
                                        className={classes.baseButton}
@@ -85,37 +99,37 @@ class LaunchBarComp extends React.Component<any,any>{
                             })}
                         </Scrollbars>
                         <div className={classes.controlPanelContainer}>
-                            <Button
+                            <Fab
                                 className={cx(classes.controlBtn, classes.controlBtnInfo)}
-                                variant={'fab'} mini color={'secondary'} aria-label={'collapse'}
+                                color='secondary' aria-label={'collapse'}
                                 onClick={handleToggleCollapse}
                             >
                                 {collapse?
                                     <ArrowForwardIcon/>:
                                     <ArrowBackIcon/>
                                 }
-                            </Button>
-                            <Button
+                            </Fab>
+                            <Fab
                                 className={cx(classes.controlBtn, classes.controlBtnSuccess)}
-                                variant={'fab'} mini color={'secondary'} aria-label={'minimize'}
+                                color='secondary' aria-label={'minimize'}
                                 onClick={handleMinimize}
                             >
                                 <RemoveIcon/>
-                            </Button>
-                            <Button
+                            </Fab>
+                            <Fab
                                 className={cx(classes.controlBtn, classes.controlBtnWarning)}
-                                variant={'fab'} mini color={'secondary'} aria-label={'expand'}
+                                color='secondary' aria-label={'expand'}
                                 onClick={handleSwitchToMainWindow}
                             >
                                 <ZoomOutMapIcon/>
-                            </Button>
-                            <Button
+                            </Fab>
+                            <Fab
                                 className={cx(classes.controlBtn, classes.controlBtnDanger)}
-                                variant={'fab'} mini color={'secondary'} aria-label={'close'}
+                                color='secondary' aria-label={'close'}
                                 onClick={handleClose}
                             >
                                 <ClearIcon/>
-                            </Button>
+                            </Fab>
                         </div>
                     </Toolbar>
                 </AppBar>
@@ -125,14 +139,14 @@ class LaunchBarComp extends React.Component<any,any>{
 }
 
 export default connect(
-    (state:any)=>({
+    (state:IRootState)=>({
         launchBarCollapse:state.application.launchBarCollapse,
     }),
     dispatch => ({
         actions:{
-            handleLaunchBarItemBtnClick:(appJson)=>()=>{dispatch(Window.actions.newWindow(appJson))},
-            handleSwitchToMainWindow:()=>{dispatch(applicationLaunchBarToggle({}))},
-            handleToggleCollapse:()=>{dispatch(applicationLaunchBarToggleCollapse({}))},
+            handleLaunchBarItemBtnClick:(appJson)=>()=>{dispatch(applicationLaunchNewWindow(appJson))},
+            handleSwitchToMainWindow:()=>{dispatch(applicationLaunchBarToggle())},
+            handleToggleCollapse:()=>{dispatch(applicationLaunchBarToggleCollapse())},
             handleMinimize:()=>{dispatch(Window.actions.minimize({}))},
             handleClose:()=>{dispatch(Window.actions.close({force:false}))},
         }
