@@ -1,13 +1,45 @@
 import * as React from 'react';
-import * as enzyme from 'enzyme';
+import { createShallow, createMount, createRender } from '@material-ui/core/test-utils';
+import Button from '@material-ui/core/Button';
+import configurestore from 'redux-mock-store';
+import { MemoryRouter } from 'react-router-dom';
+import toJson from 'enzyme-to-json';
+
+import { rootDefaultState } from './reduxs';
 
 import App from './App';
 
+const mockStore = configurestore();
+const store = mockStore(rootDefaultState);
 
-describe('App Component',()=>{
+describe('App entry',()=>{
 
-  it ('render without crashing', ()=>{
-    let wrapper:any = enzyme.shallow(<App></App>);
-    expect(wrapper).toBeTruthy();
-  })
+    let shallow;
+    let mount;
+    let render;
+
+    beforeAll(() => {
+        mount = createMount();
+        render = createRender();
+    });
+
+    afterAll(() => {
+        mount.cleanUp();
+    });
+
+
+    beforeEach(() => {
+        shallow = createShallow();
+    });
+
+    it('it renders correctly by default',()=>{
+        const wrapper = render(
+            <MemoryRouter initialEntries={['/']}>
+                <App store={store} location={{pathname:'/'}}
+                />
+            </MemoryRouter>
+        )
+        expect(wrapper).toBeTruthy();
+    })
+
 });
