@@ -31,7 +31,7 @@ const card = {
 
 const defaultFont = {
     fontFamily: '"Roboto", "Helvetica", "Arial", sans-serif',
-    fontWeight: "300",
+    fontWeight: 300,
     lineHeight: "1.5em"
 };
 
@@ -155,7 +155,7 @@ const cardLink = {
 };
 
 export {
-    //variables
+    // variables
     drawerWidth,
     appbarHeight,
     transition,
