@@ -1,9 +1,11 @@
 import * as React from 'react';
-import * as cx from 'classnames';
+import { MouseEventHandler } from "react";
+import cx from 'classnames';
 
-import { withStyles } from '@material-ui/core';
-import Button from '@material-ui/core/Button';
+import { WithStyles, withStyles } from '@material-ui/core';
+import Fab from '@material-ui/core/Fab';
 
+import AllOutIcon from '@material-ui/icons/AllOut';
 import Remove from '@material-ui/icons/Remove';
 import CropDin from '@material-ui/icons/CropDin';
 import Clear from '@material-ui/icons/Clear';
@@ -12,57 +14,69 @@ import OpenInNew from '@material-ui/icons/OpenInNew';
 
 import { headerLinksCompStyle as style } from '../../assets/jss/openfin-starter'
 
+interface IProps extends WithStyles<typeof style> {
+    windowsState:string,
+    docked?:boolean,
+    onUndock?:MouseEventHandler<any>,
+    onSwitchToLaunchBar?:MouseEventHandler<any>,
+    onMinimize:MouseEventHandler<any>,
+    onMaximize:MouseEventHandler<any>,
+    onClose:MouseEventHandler<any>,
+}
 
-class HeaderLinksComp extends React.Component<any,any>{
+class HeaderLinksComp extends React.Component<IProps,{}>{
 
     render(){
 
         const {
-            classes, windowState,
+            classes, windowsState, docked,
             onSwitchToLaunchBar,
+            onUndock,
             onMinimize, onMaximize, onClose
         } = this.props;
 
         return (<React.Fragment>
             {
                 onSwitchToLaunchBar?
-                    <Button className={cx(classes.menuBtn,classes.info)}
-                            variant="fab"
-                            mini
+                    <Fab className={cx(classes.menuBtn,classes.info)}
                             color="inherit"
                             aria-label="Switch to launch bar"
                             onClick={onSwitchToLaunchBar}
                     >
                         <OpenInNew/>
-                    </Button>:null
+                    </Fab>:null
             }
-            <Button className={cx(classes.menuBtn,classes.success)}
-                    variant="fab"
-                    mini
+            {
+                docked && onUndock ?
+                    <Fab className={cx(classes.menuBtn,classes.rose)}
+                            color="inherit"
+                            aria-label="Undock from the group"
+                            onClick={onUndock}
+                    >
+                        <AllOutIcon/>
+                    </Fab>:null
+            }
+            <Fab className={cx(classes.menuBtn,classes.success)}
                     color="inherit"
                     aria-label="Minimize"
                     onClick={onMinimize}
             >
                 <Remove/>
-            </Button>
-            <Button className={cx(classes.menuBtn,classes.warning)}
-                    variant="fab"
-                    mini
+            </Fab>
+            <Fab className={cx(classes.menuBtn,classes.warning)}
                     color="inherit"
                     aria-label="Maximize"
                     onClick={onMaximize}
             >
-                {windowState==='normal'?<CropDin/>:<AspectRatio/>}
-            </Button>
-            <Button className={cx(classes.menuBtn,classes.danger)}
-                    variant="fab"
-                    mini
+                {windowsState==='normal'?<CropDin/>:<AspectRatio/>}
+            </Fab>
+            <Fab className={cx(classes.menuBtn,classes.danger)}
                     color="inherit"
                     aria-label="Close"
                     onClick={onClose}
             >
                 <Clear/>
-            </Button>
+            </Fab>
         </React.Fragment>);
     }
 
