@@ -1,8 +1,9 @@
 import * as React from 'react';
-import * as cx from 'classnames';
+import { MouseEventHandler } from "react";
+import cx from 'classnames';
 
-import Snackbar from '@material-ui/core/Snackbar';
 import SnackbarContent from '@material-ui/core/SnackbarContent';
+import {SnackbarContentProps} from "@material-ui/core/SnackbarContent/SnackbarContent";
 import IconButton from '@material-ui/core/IconButton';
 
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
@@ -11,8 +12,9 @@ import ErrorIcon from '@material-ui/icons/Error';
 import InfoIcon from '@material-ui/icons/Info';
 import CloseIcon from '@material-ui/icons/Close';
 import WarningIcon from '@material-ui/icons/Warning';
+import FaceIcon from '@material-ui/icons/Face'
 
-import {withStyles} from '@material-ui/core/styles';
+import { WithStyles, withStyles} from '@material-ui/core/styles';
 import { snackbarContentCompStyle as style } from '../../assets/jss/openfin-starter';
 
 const variantIcon ={
@@ -21,9 +23,20 @@ const variantIcon ={
     warning: WarningIcon,
     error: ErrorIcon,
     info: InfoIcon,
+    rose: FaceIcon,
 };
 
-class MySnackbarContentComp extends React.Component<any,any>{
+interface IMySnackbarContentProps{
+    classes?:any,
+    className?:string,
+    message:string,
+    onClose:MouseEventHandler<any>,
+    variant:string,
+}
+
+type Props = IMySnackbarContentProps & WithStyles<typeof style> & SnackbarContentProps
+
+class MySnackbarContentComp extends React.Component<Props,{}>{
     render(){
 
         const {
