@@ -1,51 +1,65 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { BrowserAdapter } from '@albertli/openfin-browser-adapter';
+import { BrowserAdapter } from '@albertli90/openfin-browser-adapter';
 
 import './assets/css/main.css';
 
 import setPlatformClass from './utils/setPlatformClass';
 import App from './App';
-import registerServiceWorker from './registerServiceWorker';
-import {applicationStarted} from "./redux";
+import * as serviceWorker from './serviceWorker';
 
 import configureStore from './utils/configureStore';
+import {
+    CLIENT_SET_VALUE,
+    applicationStarted,
+    applicationChildStarted,
+    applicationNetworkOnline,
+    applicationNetworkOffline,
+} from "./reduxs";
 
 declare const window:any;
 
+const sharedActions = [
+    CLIENT_SET_VALUE
+];
+
 if(!window.fin){
-    window.fin = new BrowserAdapter({userSocket:false});
+    window.fin = new BrowserAdapter({
+        finUuic:process.env.REACT_APP_FIN_UUID,
+        silentMode:false,
+    });
 }
 
-if (process.env.REACT_APP_SHARED_REDUX_STORE === 'true'){
-    if(window.store == null && window.opener == null){
-        const store = configureStore();
-        window.store=store;
-        store.dispatch(applicationStarted());
-    }
-    setPlatformClass(document.body,window.navigator.platform);
-    ReactDOM.render(
-        <Provider store = {window.opener?window.opener.store:window.store}>
-            <App/>
-        </Provider>
-        ,
-        document.getElementById('root')
+if(window.name === process.env.REACT_APP_FIN_UUID){
+    const store = configureStore(
+        sharedActions,
     );
-}else{
-    const store = configureStore();
     window.store=store;
     store.dispatch(applicationStarted());
-
-    setPlatformClass(document.body,window.navigator.platform);
-
-    ReactDOM.render(
-        <Provider store = {window.store}>
-            <App/>
-        </Provider>
-        ,
-        document.getElementById('root')
+}else{
+    const store = configureStore(
+        sharedActions,
+        window.opener.store.getState()
     );
+    window.store=store;
+    store.dispatch(applicationChildStarted());
 }
+setPlatformClass(document.body,window.navigator.platform);
+ReactDOM.render(
+    <Provider store = {window.store}>
+        <App/>
+    </Provider>
+    ,
+    document.getElementById('root')
+);
 
-registerServiceWorker();
+serviceWorker.unregister();
+
+
+window.addEventListener('online',()=>{
+    window.store.dispatch(applicationNetworkOnline());
+})
+window.addEventListener('offline',()=>{
+    window.store.dispatch(applicationNetworkOffline());
+})
