@@ -1,24 +1,27 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Switch, Route, Redirect } from 'react-router-dom';
-import * as cx from 'classnames';
-
+import cx from 'classnames';
 import Snackbar from '@material-ui/core/Snackbar';
-import IconButton from '@material-ui/core/IconButton';
 
-import CloseIcon from '@material-ui/icons/Close';
+import { WithStyles ,withStyles} from '@material-ui/core/styles';
 
-import {withStyles} from '@material-ui/core/styles';
+import {
+    // acitons
+    applicationToogleWindowState,
+    applicationSetSnackbarStatus, applicationProcessSnackbarQueue,
+    applicationCloseSnackbar,
+    // types
+    IRootState,ISnackBarMsg,
+} from '../../reduxs';
 
-import { Window } from '@albertli/redux-openfin';
 
-import { RouteItem, RouteCompItem, RouteRedirectItem } from '../../routes';
 import childWindowRoutes from '../../routes/ChildWindow';
 
-import {Header} from '../../components';
+import { Header,SnackbarContent } from '../../components';
 
 import { dashboardLayoutStyle as style } from '../../assets/jss/openfin-starter';
-import dashboardRoutes from "../../routes/Dashboard";
+import {Window} from "@albertli90/redux-openfin";
 
 declare const window:any;
 
@@ -33,72 +36,55 @@ const switchRoutes = (
     </Switch>
 );
 
-class ChildWindowLayout extends React.Component<any,any>{
+interface IProps extends WithStyles<typeof style>{
+    docked:boolean,
+    snackBarOpen:boolean,
+    snackBarMsgInfo:Partial<ISnackBarMsg>,
+    windowsState:string,
+    actions:{
+        handleSetAsForeground: ()=> void,
+        handleSnackbarClose: (event:any,reason:string)=> void,
+        handleSnackbarCloseBtnClick: ()=> void,
+        handleSnackbarExited: ()=> void,
+        handleUndock: ()=> void,
+        handleMinimize: ()=> void,
+        handleMaximize: ()=> void,
+        handleClose: ()=> void,
+    }
+}
 
-    state={
-        windowState:'normal',
-    };
+class ChildWindowLayout extends React.Component<IProps,{}>{
 
     componentDidMount(){
-        this.getCurrentWindow().setAsForeground();
+        this.props.actions.handleSetAsForeground();
     }
-
-    getCurrentWindow= ()=>{
-        return window.fin.desktop.Window.getCurrent();
-    };
-
-    updateWindowState = ()=>{
-        this.getCurrentWindow().getState(
-            (state)=>{
-                this.setState({windowState:state});
-            }
-        );
-    };
-
-    handleMinimize = ()=>{
-        this.getCurrentWindow().minimize();
-        this.updateWindowState();
-    };
-
-    handleMaximize = ()=>{
-        const { windowState } = this.state;
-        if (windowState === 'maximized'){
-            this.getCurrentWindow().restore();
-        }else if (windowState === 'normal'){
-            this.getCurrentWindow().maximize();
-        }
-        this.updateWindowState();
-    };
-
-    handleClose = () =>{
-        this.getCurrentWindow().close();
-    };
 
     render(){
         const {
             classes,
+            docked,snackBarOpen, snackBarMsgInfo, windowsState,
             actions:{
-
+                handleSetAsForeground,
+                handleSnackbarClose, handleSnackbarCloseBtnClick, handleSnackbarExited,
+                handleUndock, handleMinimize, handleMaximize, handleClose,
             },
             ...rest
         } = this.props;
-
-        const {windowState} = this.state;
-
-        this.updateWindowState();
 
         return(
             <React.Fragment>
                 <Header
                     routes={childWindowRoutes}
-                    windowsState={windowState}
+                    windowsState={windowsState}
                     color={'info'}
-                    onMinimize={this.handleMinimize}
-                    onMaximize={this.handleMaximize}
-                    onClose = {this.handleClose}
+                    docked={docked}
+                    onUndock = {handleUndock}
+                    onMinimize={handleMinimize}
+                    onMaximize={handleMaximize}
+                    onClose = {handleClose}
                     {...rest}
                 />
-                <div className={cx(classes.wrapper, classes.wrapperPrimary)}>
+                <div className={cx(classes.wrapper, 'info-top-to-bottom')}>
                     <div className={cx(
                         classes.mainPanel,classes.mainPanelShift
                     )}
@@ -110,6 +96,23 @@ class ChildWindowLayout extends React.Component<any,any>{
                         </div>
                     </div>
                 </div>
+                <Snackbar
+                    key={snackBarMsgInfo.key}
+                    anchorOrigin={{
+                        vertical:'bottom',
+                        horizontal:'center'
+                    }}
+                    open={snackBarOpen}
+                    autoHideDuration={6000}
+                    onClose={handleSnackbarClose}
+                    onExited={handleSnackbarExited}
+                >
+                    <SnackbarContent
+                        onClose={handleSnackbarCloseBtnClick}
+                        variant={snackBarMsgInfo.variant}
+                        message={snackBarMsgInfo.message}
+                    />
+                </Snackbar>
             </React.Fragment>
         );
     }
@@ -117,12 +120,23 @@ class ChildWindowLayout extends React.Component<any,any>{
 }
 
 export default connect(
-    (state:any) => ({
-
+    (state:IRootState) => ({
+        docked:state.application.docked,
+        snackBarOpen:state.application.snackBarOpen,
+        snackBarMsgInfo:state.application.snackBarMsgInfo,
+        windowsState:state.application.windowsState,
     }),
     dispatch => ({
         actions:{
-
+            handleSetAsForeground: () => {dispatch(Window.actions.setAsForeground({}))},
+            handleSnackbarClose: (event,reason) => {dispatch(applicationCloseSnackbar({event,reason}))},
+            handleSnackbarCloseBtnClick: () => {dispatch(applicationSetSnackbarStatus({open:false}))},
+            handleSnackbarExited: () => {dispatch(applicationProcessSnackbarQueue())},
+            // openfin
+            handleUndock : () => {dispatch(Window.actions.leaveGroup({}))},
+            handleMinimize: ()=>{dispatch(Window.actions.minimize({}))},
+            handleMaximize: ()=>{dispatch(applicationToogleWindowState())},
+            handleClose:()=>{dispatch(Window.actions.close({force:false}))},
         }
     })
 )(withStyles(style)(ChildWindowLayout));
