@@ -1,17 +1,28 @@
 import { applyMiddleware, createStore, compose } from 'redux';
 import createSagaMiddleware from 'redux-saga';
-import { createOpenfinMiddleware } from '@albertli/redux-openfin';
+import { createOpenfinMiddleware } from '@albertli90/redux-openfin';
+import { ChannelType } from '@albertli90/redux-openfin/init';
 
-import rootReducer from '../redux';
-import rootSaga from '../redux/sagas';
+import rootReducer, {IRootState} from '../reduxs';
+import rootSaga from '../reduxs/sagas';
 
 declare const window:any;
 
-export default ()=>{
+export default (
+        sharedActions:string[],
+        parentState?:IRootState
+)=>{
 
-    const openfinMiddleware = createOpenfinMiddleware(window.fin);
+    const openfinMiddleware = createOpenfinMiddleware(window.fin,{
+        finUuid:process.env.REACT_APP_FIN_UUID,
+        sharedActions,
+        // channelRandomSuffix:process.env.NODE_ENV === 'development',
+        autoDocking:process.env.REACT_APP_ENABLE_AUTO_DOCKING === 'true',
+        dockingOptions:{
+        }
+    });
     const sagaMiddleware = createSagaMiddleware();
-    const devtools = window['devToolsExtension']?window['devToolsExtension']():(f:any):any => (f);
+    const devtools = window.devToolsExtension?window.devToolsExtension():(f:any):any => (f);
 
     const middleware = compose(
         applyMiddleware(
@@ -22,7 +33,7 @@ export default ()=>{
     );
 
     const store = createStore(
-        rootReducer,
+        rootReducer(parentState),
         middleware,
     );
 
